修复： 1.GUI按下shift点击空白的输入框会导致程序崩溃 更改： 1.重写了合成表 2.优化了导入 -------------------- 已知问题： 1.（偶然）机器在电线被破坏后仍然可以接收能量 2.发电机输出电能的同时发电会有电能计算错误

+++ src/main/java/xyz/emptydreams/mi/api/list/IntegerList.java
@@ -129,7 +129,7 @@ public class IntegerList implements List<Integer>, RandomAccess {
 	public boolean containsAll(Collection<?> c) {
 		for (Object o : c) {
 			if (o.getClass() == Integer.class) {
-				if (!contains((Integer) o)) return false;
+				if (!contains(o)) return false;
 			} else {
 				return false;
 			}
